simplified extractors files
